requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
djangorestframework>=3.14
django-filter>=24.2
django-storages>=1.14
boto3>=1.34
Pillow>=10.0
qrcode[pil]>=7.4
python-dateutil>=2.8
icalendar>=5.0
crispy-bootstrap5>=2023.10
django-crispy-forms>=2.1
requests>=2.31
